+++ src/core/operations/Shake.mjs
@@ -20,7 +20,7 @@ class Shake extends Operation {
         super();
 
         this.name = "Shake";
-        this.module = "Hashing";
+        this.module = "Crypto";
         this.description = "Shake is an Extendable Output Function (XOF) of the SHA-3 hash algorithm, part of the Keccak family, allowing for variable output length/size.";
         this.infoURL = "https://wikipedia.org/wiki/SHA-3#Instances";
         this.inputType = "ArrayBuffer";